+++ hello.py
@@ -1,1 +1,1 @@
-print("hello from traceforge")
+print("changed by traceforge command")